fixed dropdown menu, error message, & search bar
removed "FED 0" from dropdown menu, as it was just a placeholder.

Also changed width to 100% rather than 50vw so it matched the search bar at all times.

The search bar now gets a new ID on every run so that most browsers don't try to show suggestions for it even after ignoring the autocomplete="off".